+++ app/routes/components/Tabs/changelog.codegen.ts
@@ -5,4 +5,6 @@ export const changelog = ''
   + '- 2025-10-05: Initial version.\n'
   + '- 2025-10-16: Adjust its styles - improve spacing and hover states.\n'
   + '- 2025-10-31: Support the `enclosed` variant.\n'
+  + '- 2025-11-15: Adjust spacing between tabs for better visual balance. Remove the\n'
+  + '  dependency on deprecated `ThemeProvider` and use `CopyUiProvider` instead.\n'
   ;